backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
PyPDF2==3.0.1
python-multipart==0.0.6
firecrawl==2.15.0
python-dotenv==1.0.0
supabase==2.3.4
openai==1.12.0
